Moved grid resources to predator prey example
