--- admiral/envs/modules/__init__.py
@@ -1,1 +1,0 @@
-from .grid_resources import GridResources
